server/app/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.0
Django==3.2.8
django-cors-headers==3.10.0
django-mysql==4.2.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
elasticsearch==7.15.2
gunicorn==20.1.0
idna==2.7
isodate==0.6.0
msrest==0.6.21
mysqlclient==1.4.0
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.2.0
pytz==2021.3
requests==2.20.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.24.3
